requirements.txt
crewai
PyPDF2
crewai_tools
openai
python-dotenv

langchain_google_genai
google.generativeai